requirements.txt
Pillow==9.5.0
pystray==0.19.4
Requests==2.31.0
